% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imtp-package.R
\docType{package}
\name{imtp-package}
\alias{imtp}
\alias{imtp-package}
\title{imtp: What the Package Does (Title Case)}
\description{
...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nt-williams/imtp}
  \item Report bugs at \url{https://github.com/nt-williams/imtp/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Williams \email{ntwilliams.personal@gmail.com} (\href{https://orcid.org/0000-0002-1378-4831}{ORCID}) [copyright holder]

}
\keyword{internal}
