% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imtp_simul.R
\name{imtp_simul}
\alias{imtp_simul}
\title{Estimate simultaneous confidence intervals}
\usage{
imtp_simul(fits, reps = 1e+05, level = 0.95)
}
\arguments{
\item{fits}{[\code{list}]\cr
A list of \code{imtp} fits.}

\item{reps}{[\code{numeric(1)}]\cr
The number of repitions to use for the multiplier bootstrap. The default is 1e5.}

\item{level}{[\code{numeric(1)}]\cr
The desired confidence level. The default is 0.95.}
}
\value{
A \code{data.frame} with estimates.
}
\description{
Estimate simultaneous confidence intervals
}
\examples{
n <- 1000
W <- matrix(rnorm(n*3), ncol = 3)
A <- rbinom(n, 1, 1/(1 + exp(-(.2*W[,1] - .1*W[,2] + .4*W[,3]))))
Y <- A + 2*W[,1] + W[,3] + W[,2]^2 + rnorm(n)
R <- rbinom(n, 1, 0.9)
tmp <- data.frame(W, A, R, Y = ifelse(R == 1, Y, NA_real_))
deltas <- seq(0.1, 2, length.out = 5)
fits <- lapply(deltas, function(d) imtp_tmle(tmp, "A", "Y", paste0("X", 1:3), cens = "R", folds = 1, delta = d, outcome_type = "continuous"))
imtp_simul(fits)
}
