% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{imtp_tmle}
\alias{imtp_tmle}
\title{Incremental Propensity Score Intervention Targeted Maximum Likelihood Estimator}
\usage{
imtp_tmle(
  data,
  trt,
  outcome,
  baseline = NULL,
  time_vary = NULL,
  cens = NULL,
  delta = 1,
  k = Inf,
  mtp = FALSE,
  outcome_type = c("binomial", "continuous", "survival"),
  id = NULL,
  bounds = NULL,
  learners_outcome = "SL.glm",
  learners_trt = "SL.glm",
  folds = 10,
  weights = NULL,
  .bound = 1e-05,
  .trim = 0.999,
  .learners_outcome_folds = 10,
  .learners_trt_folds = 10,
  .return_full_fits = FALSE,
  ...
)
}
\arguments{
\item{data}{\[\code{data.frame}\]\cr
A \code{data.frame} in wide format containing all necessary variables
for the estimation problem. Must not be a \code{data.table}.}

\item{trt}{\[\code{character}\]\cr
A vector containing the column names of treatment variables ordered by time.}

\item{outcome}{\[\code{character}\]\cr
The column name of the outcome variable. In the case of time-to-event
analysis, a vector containing the columns names of intermediate outcome variables and the final
outcome variable ordered by time. Only numeric values are allowed. If the outcome type
is binary, data should be coded as 0 and 1.}

\item{baseline}{\[\code{character}\]\cr
An optional vector containing the column names of baseline covariates to be
included for adjustment at every time point.}

\item{time_vary}{\[\code{list}\]\cr
A list the same length as the number of time points of observation with
the column names for new time-varying covariates introduced at each time point. The list
should be ordered following the time ordering of the model.}

\item{cens}{\[\code{character}\]\cr
An optional vector of column names of censoring indicators the same
length as the number of time points of observation. If missingness in the outcome is
present or if time-to-event outcome, must be provided.}

\item{delta}{\[\code{numeric(1)}\]\cr}

\item{k}{\[\code{integer(1)}\]\cr
An integer specifying how previous time points should be
used for estimation at the given time point. Default is \code{Inf},
all time points.}

\item{outcome_type}{\[\code{character(1)}\]\cr
Outcome variable type (i.e., continuous, binomial, survival).}

\item{id}{\[\code{character(1)}\]\cr
An optional column name containing cluster level identifiers.}

\item{bounds}{\[\code{numeric(2)}\]\cr
An optional, ordered vector of the bounds for a continuous outcomes. If \code{NULL},
the bounds will be taken as the minimum and maximum of the observed data.
Should be left as \code{NULL} if the outcome type is binary.}

\item{learners_outcome}{\[\code{character}\]\cr A vector of \code{SuperLearner} algorithms for estimation
of the outcome regression. Default is \code{"SL.glm"}, a main effects GLM.}

\item{learners_trt}{\[\code{character}\]\cr A vector of \code{SuperLearner} algorithms for estimation
of the exposure mechanism. Default is \code{"SL.glm"}, a main effects GLM.}

\item{folds}{\[\code{integer(1)}\]\cr
The number of folds to be used for cross-fitting.}

\item{weights}{\[\code{numeric(nrow(data))}\]\cr
An optional vector containing sampling weights.}

\item{.bound}{\[\code{numeric(1)}\]\cr
Determines that maximum and minimum values (scaled) predictions
will be bounded by. The default is 1e-5, bounding predictions by 1e-5 and 0.9999.}

\item{.trim}{\[\code{numeric(1)}\]\cr
Determines the amount the density ratios should be trimmed.
The default is 0.999, trimming the density ratios greater than the 0.999 percentile
to the 0.999 percentile. A value of 1 indicates no trimming.}

\item{.learners_outcome_folds}{\[\code{integer(1)}\]\cr
The number of cross-validation folds for \code{learners_outcome}.}

\item{.learners_trt_folds}{\[\code{integer(1)}\]\cr
The number of cross-validation folds for \code{learners_trt}.}

\item{.return_full_fits}{\[\code{logical(1)}\]\cr
Return full SuperLearner fits? Default is \code{FALSE}, return only SuperLearner weights.}

\item{...}{Extra arguments. Exists for backwards compatibility.}
}
\value{
A list of class \code{imtp} containing the following components:

\item{estimator}{The estimator used, in this case "TMLE".}
\item{theta}{The estimated population effect.}
\item{standard_error}{The estimated standard error of the estimand.}
\item{low}{Lower bound of the 95\% confidence interval of the estimand.}
\item{high}{Upper bound of the 95\% confidence interval of the estimand.}
\item{eif}{The estimated, un-centered, influence function of the estimate.}
\item{delta}{...}
\item{outcome_reg}{An n x Tau + 1 matrix of outcome regression predictions.
 The mean of the first column is used for calculating theta.}
\item{density_ratios}{An n x Tau matrix of the estimated, non-cumulative, density ratios.}
\item{fits_m}{A list the same length as \code{folds}, containing the fits at each time-point
 for each fold for the outcome regression.}
\item{fits_r}{A list the same length as \code{folds}, containing the fits at each time-point
for each fold of density ratio estimation.}
\item{outcome_type}{The outcome variable type.}
}
\description{
Incremental Propensity Score Intervention Targeted Maximum Likelihood Estimator
}
\examples{

}
